refactor: Calculator 분리
- 당첨 결과 산정 로직을 LottoCenter 객체에 위임
- 구매한 수량을 계산하는 메서드 추가
- 수량 계산 테스트 추가
- 객체 분리에 따라 테스트도 분리
